requirements.txt
fastapi
uvicorn
jinja2
requests
python-multipart
firebase-admin
pandas
pyserial
websockets
schedule
openmeteo_requests
requests_cache
retry_requests
